test version 1 with flask server
