+++ requirements.txt
@@ -1,4 +1,4 @@
 discord.py==2.0.1
 python-dotenv==0.20.0
 openai
-firebase_functions
+flask
